third_party/blink/web_tests/http/tests/devtools/sources/debugger-breakpoints/set-logpoint-expected.txt

Tests setting logpoints.


Running: testSetLogpoint
Script source was shown.
Set timer for test function.
Test function finished.
Breakpoint sidebar pane 
set-breakpoint.html:15  return x;
Message count: 1
VM:1 x is 3.1622776601683795
Breakpoints removed.
Breakpoint sidebar pane 
No breakpoints
